requirements.txt
fastapi[all]~=0.111.0
python-dotenv~=1.0.1
pydantic~=2.8.2
requests~=2.32.3